Adapted ROadRunnerPy to work with latest version of libroadrunner - the changes included replacing rr.simulateOptions.relative with rr.getIntegrator().relative_tolerance=... rr.getIntegrator().max_num_steps etc...
There is one reference to rr.options in SBMLHelper but thiswill be fixed late once I figure out correct way of handling it.

Note: some of the older options , notably steps (relabeled to max_num_steps in the latestrr) need to be changed in the example code. In particular old number of steps 10 is often insufficient so thechoice is to either remove it completely or set to a much higher number